fixed inconsistent pDop type

+++ software/src/fmu/sensors.h
@@ -186,7 +186,7 @@ class uBloxSensor {
       float HorizontalAccuracy_m;               // Accuracy Horizontal (m)
       float VerticalAccuracy_m;                 // Accuracy Vertical (m)
       float VelocityAccuracy_ms;                // Accuracy Speed (m/s)
-      double pDOP;                              // Position DOP
+      float pDOP;                               // Position DOP
     };
     void UpdateConfig(const char *JsonString,std::string RootPath,DefinitionTree *DefinitionTreePtr);
     void SetConfig(const Config &ConfigRef);
