+++ software/src/node/sensors.h
@@ -143,7 +143,7 @@ class uBloxSensor {
       float HorizontalAccuracy_m;               // Accuracy Horizontal (m)
       float VerticalAccuracy_m;                 // Accuracy Vertical (m)
       float VelocityAccuracy_ms;                // Accuracy Speed (m/s)
-      double pDOP;                              // Position DOP
+      float pDOP;                              // Position DOP
     };
     void UpdateConfig(const char *JsonString);
     void SetConfig(const Config &ConfigRef);
